✅[CHORE] : 레이아웃 코드 분리 (#20) - SectionLayout 파일로 레이아웃 코드 분리 - Enum CaseIterable 채택으로 케이스 명확히 분리 (default값 미사용) - 접근 제어자 지정 (private) - 불필요 코드 제거
