+++ 29thAssignment/29thAssignment/Views/Home/HomeViewController.swift
@@ -9,29 +9,19 @@ import UIKit
 
 import SnapKit
 import Then
-import SwiftUI
-
-// MARK: - HomeVC에서 사용될 셀 타입
-
-private enum CellType {
-    case channel
-    case category
-    case feed
-}
 
 class HomeViewController: UIViewController {
     
     // MARK: - UI properties
-    lazy var collectionView = UICollectionView(
+    private lazy var collectionView = UICollectionView(
         frame: .zero,
         collectionViewLayout: createLayout()
     )
-    var tableView = UITableView()
+    private var tableView = UITableView()
     
     // MARK: - Properties
-    fileprivate let layout: [CellType] = [.channel, .category, .feed]
     private let channels: [Channel] = Channel.loadDummyList()
-    private let categoryTitles = [ "전체", "오늘", "이어서 시청하기", "시청하지 않음", "실시간", "게시물"]
+    private let categoryTitles = ["전체", "오늘", "이어서 시청하기", "시청하지 않음", "실시간", "게시물"]
     private let feeds: [Feed] = Feed.loadDummyList()
 
     // MARK: - Life cycles
@@ -90,96 +80,26 @@ class HomeViewController: UIViewController {
     }
 }
 
-extension HomeViewController {
-    // MARK: - Composional Layout
-    
-    func createLayout() -> UICollectionViewCompositionalLayout {
-        UICollectionViewCompositionalLayout { (sectionIndex, env)
-            -> NSCollectionLayoutSection? in
-            switch sectionIndex {
-            case 0: return self.createChannelSection()
-            case 1: return self.createCategorySection()
-            case 2: return self.createFeedSection(env)
-            default: return nil
-            }
-        }
-    }
-    
-    fileprivate func supplementaryFooterSeparatorItem() -> NSCollectionLayoutBoundarySupplementaryItem {
-        NSCollectionLayoutBoundarySupplementaryItem(
-            layoutSize: NSCollectionLayoutSize(
-                widthDimension: .fractionalWidth(1.0),
-                heightDimension: .absolute(1)),
-            elementKind: UICollectionView.elementKindSectionFooter,
-            alignment: .bottom)
-    }
-    
-    private func createChannelSection() -> NSCollectionLayoutSection {
-        let item = NSCollectionLayoutItem(
-            layoutSize: .init(
-                widthDimension: .estimated(1),
-                heightDimension: .estimated(1)
-            )
-        )
-        let group = NSCollectionLayoutGroup.horizontal(
-            layoutSize: .init(
-                widthDimension: .absolute(72),
-                heightDimension: .absolute(104)
-            ), subitems: [item]
-        )
-        let section = NSCollectionLayoutSection(group: group)
-        section.orthogonalScrollingBehavior = .continuous
-        section.supplementariesFollowContentInsets = false
-        section.boundarySupplementaryItems = [supplementaryFooterSeparatorItem()]
-        return section
-    }
-    
-    private func createCategorySection() -> NSCollectionLayoutSection {
-        let item = NSCollectionLayoutItem(
-            layoutSize: .init(
-                widthDimension: .estimated(1),
-                heightDimension: .absolute(32)
-            )
-        )
-        let group = NSCollectionLayoutGroup.horizontal(
-            layoutSize: .init(
-                widthDimension: .estimated(1),
-                heightDimension: .absolute(48)
-            ), subitems: [item]
-        )
-        let section = NSCollectionLayoutSection(group: group)
-        section.interGroupSpacing = 9
-        section.contentInsets = .init(top: 9, leading: 13, bottom: -9, trailing: 13)
-        section.orthogonalScrollingBehavior = .continuous
-        return section
-    }
-    
-    private func createFeedSection(_ env: NSCollectionLayoutEnvironment) -> NSCollectionLayoutSection {
-        let listConfig = UICollectionLayoutListConfiguration(appearance: .plain)
-        return NSCollectionLayoutSection.list(using: listConfig,layoutEnvironment: env)
-    }
-}
-
 // MARK: - CollectionView DataSource
 
 extension HomeViewController: UICollectionViewDataSource {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
-        layout.count
+        SectionLayout.allCases.count
     }
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        switch layout[section] {
+        switch SectionLayout.allCases[section] {
         case .channel:
             return channels.count
         case .category:
             return categoryTitles.count
-        default:
+        case .feed:
             return feeds.count
         }
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        switch layout[indexPath.section] {
+        switch SectionLayout.allCases[indexPath.section] {
         case .channel:
             let cell = collectionView.dequeueReusableCell(forIndexPath: indexPath) as ChannelCollectionViewCell
             let channel = channels[indexPath.row]
@@ -190,7 +110,7 @@ extension HomeViewController: UICollectionViewDataSource {
             let cell = collectionView.dequeueReusableCell(forIndexPath: indexPath) as CategoryCollectionViewCell
             cell.categoryTitleLabel.text = categoryTitles[indexPath.row]
             return cell
-        default:
+        case .feed:
             let cell = collectionView.dequeueReusableCell(forIndexPath: indexPath) as FeedCollectionViewCell
             let feed = feeds[indexPath.row]
             cell.configure(thumbnailImage: feed.thumbnailImage,
